Got all of the scenes except for Evening/Morning/Final Scores fixed at 800x600
Decided on 800x600 as it fitted reasonable within browser window

first tried to allow scaling - but it's an absolute arse on to get it working with the UI - so I decided to forget it and just put a black border around if they fullscreen the game. 800x600 ought to be enough for anybody!

Had some trouble making the scenes I wrote before which work (Evening/Morning) fit with the new 800x600 frame. I'll try to get that done now.

If I finish that then we then have a complete flow through the game (albeit with most of the "game" missing)