Moved all old code into "legacy" - started rewrite
Our code wasn't nice so I've scrapped it and rewritten the MainMenu and some of the Lobby.

Using http://www.daikonforge.com/dfgui/ for the GUI now - it looks ugly as sin (some horrible scifi theme is the default) but we can sort that out later - it's a lot easier to work with than NGUI and it has some awesome features (such as data binding)

I'll probably be able to finish the lobby today.